New tab was opening new window
Using the shift + spacebar shortcut was opening highlighted URL in a new window instead of a new tab. This is probably due to browser settings that can't be changed. I changed the shortcut to ctrl+shift+enter since control is the normal modifier for new tabs. Should work on both Windows and Mac. Haven't tested Linux yet.